Remove Link component it moves to FrouteLink
